# amp-bind: Support data-amp-bind-* attributes