Fixes #11115. Also fixes a few lint errors.

- Support `data-amp-bind-foo="bar"` as an alternative to `[foo]="bar"`.

Performance impact of additional string operations seems negligible. Time spent scanning `examples/bind/performance.amp.html` on Mac Pro 2013 with 6x CPU throttling (10 samples each):

| Before (median) | After (median) |
| -- | -- |
| 570.15 ms | 562.1 ms |

/to @aghassemi